+++ server/src/utils/sleep.ts
@@ -1,0 +1,3 @@
+import { promises } from "fs";
+
+export const sleep = (ms: number) => new Promise((res) => setTimeout(res, ms));