+++ web/src/constants.ts
@@ -1,0 +1,1 @@
+export const __containerHeight__ = "calc(100vh - 2.5rem);";